# Use godirwalk instead of filepath.Walk for better performance