[godirwalk.Walk](https://github.com/karrick/godirwalk) is faster than filepath.Walk.

It's two times faster in my environment.

Current (filepath.Walk)
```
❯ time ghq list | wc -l
      63
ghq list  0.07s user 0.15s system 99% cpu 0.218 total
wc -l  0.00s user 0.00s system 1% cpu 0.217 total
```


godirwalk
```
❯ go build -o $GOPATH/bin/my-ghq
```

```
❯ time my-ghq list | wc -l
      63
my-ghq list  0.05s user 0.08s system 96% cpu 0.130 total
wc -l  0.00s user 0.00s system 2% cpu 0.130 total
```