@Songmu 2019-04-26
Thank you!

However, if it is 10 times faster, it is worth considering, but if it is just only this difference, I would like to expect the reliability and future improvement of the standard library.

So I stay using `filepath.Walk`.